+++ NEWS.md
@@ -8,6 +8,9 @@ This can be used for adding `AVISIT`, `AWLO`, `AWHI` based on time windows and
 `ADY` or deriving the lowest value (nadir) before the current observation
 (#1448).
 
+- The new function `chr2vars()` turns a character vector into a list of quosures
+(#1448).
+
 - A `print()` method was added for all S3 objects defined by admiral, e.g.,
 `date_source()`, `dthcaus_source()`, ... (#858)